requirements.txt
flask==2.3.3
transformers==4.35.2
torch==2.1.1
sentencepiece==0.1.99
sacremoses==0.0.53
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
requests==2.31.0
flask-cors==4.0.0
python-dotenv==1.0.0
datasets==2.14.6
accelerate==0.24.1
